+++ src/Form/Element/NamedFormElement.php
@@ -407,8 +407,7 @@ abstract class NamedFormElement extends FormElement
         if ($request->hasSession() && ! is_null($value = $request->old($this->getPath()))) {
             return $value;
         }
-
-        // return $request->input($this->getPath());
+        
         return $request->input($this->getNameKey()) ?: $request->input($this->getPath());
     }
 
@@ -431,6 +430,7 @@ abstract class NamedFormElement extends FormElement
               */
             return $value;
         }
+
 
         /*
           * Implement json parsing